refactor: split ingestion operations dtos
